requirements.txt
discord.py
python-dotenv
requests
db-sqlite3
aiosqlite